Move Skirt into draft-page-layout package
