+++ draft-packages/page-layout/KaizenDraft/Skirt/Skirt.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 import { Container, Content } from "@kaizen/draft-page-layout"
 import classNames from "classnames"
+import { DOMRectReadOnly, useResizeObserver } from "../useResizeObserver"
 import styles from "./styles.scss"
-import { DOMRectReadOnly, useResizeObserver } from "./useResizeObserver"
 
 const spacing = 24
 const maxHeight = 222
