+++ draft-packages/page-layout/KaizenDraft/Skirt/SkirtCard.scss
