+++ draft-packages/page-layout/index.ts
@@ -1,1 +1,3 @@
 export * from "./KaizenDraft/Content/Content"
+export * from "./KaizenDraft/Skirt/Skirt"
+export * from "./KaizenDraft/Skirt/SkirtCard"
