+++ draft-packages/page-layout/package.json
@@ -31,7 +31,10 @@
   "private": false,
   "license": "MIT",
   "dependencies": {
-    "classnames": "^2.2.6"
+    "@kaizen/draft-card": "^1.2.4",
+    "@kaizen/draft-page-layout": "^1.0.0",
+    "classnames": "^2.2.6",
+    "resize-observer-polyfill": "^1.5.1"
   },
   "peerDependencies": {
     "@kaizen/design-tokens": "^2",
