--- draft-packages/skirt/README.mdx
@@ -1,39 +1,0 @@
----
-title: "Skirt"
-navTitle: "Skirt"
-summaryParagraph: Skirt lets the user...
-tags: ["Common alternative name", "Another example"]
-demoStoryId:  Skirt-react--default-kaizen-site-demo
----
-
-import WhenToUseAndWhenNotToUse from "docs-components/WhenToUseAndWhenNotToUse"
-import WhenToUse from "docs-components/WhenToUse"
-import WhenNotToUse from "docs-components/WhenNotToUse"
-import DoAndDontContainer from "docs-components/DoAndDontContainer"
-import Do from "docs-components/Do"
-import Dont from "docs-components/Dont"
-
-<WhenToUseAndWhenNotToUse>
-
-<WhenToUse>
-
-* Use Skirt when...
-
-</WhenToUse>
-
-<WhenNotToUse>
-
-* When X, use Y component instead.
-
-</WhenNotToUse>
-
-</WhenToUseAndWhenNotToUse>
-
-## See also
-- TODO - Is there a relevant or related component? Add it here.
-* [Button](/components/button/)
-
-## External links
-- TODO - find relevant examples from other systems and add them here.
-- [Polaris: button](https://polaris.shopify.com/components/actions/button)
-- [Atlassian: button](https://www.atlassian.design/components/button/examples)
