--- draft-packages/skirt/index.ts
@@ -1,2 +1,0 @@
-export * from "./KaizenDraft/Skirt/Skirt"
-export * from "./KaizenDraft/Skirt/SkirtCard"
