--- draft-packages/skirt/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@kaizen/draft-skirt",
-  "version": "1.0.0",
-  "description": "The draft skirt component",
-  "scripts": {
-    "prepublish": "tsc --project tsconfig.dist.json",
-    "build": "yarn clean && yarn prepublish",
-    "build:watch": "yarn clean && yarn prepublish --watch",
-    "clean": "rimraf '*.d.ts' '*.js' '*.map' 'KaizenDraft/**/*.d.ts' 'KaizenDraft/**/*.js' 'KaizenDraft/**/*.map'"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cultureamp/kaizen-design-system.git",
-    "directory": "draft-packages/skirt"
-  },
-  "bugs": {
-    "url": "https://github.com/cultureamp/kaizen-design-system/issues"
-  },
-  "files": [
-    "**/*",
-    "!**/*.ts",
-    "!**/*.tsx",
-    "**/*.d.ts",
-    "!**/*.spec.*",
-    "!**/*.snap",
-    "!draft-templates",
-    "!stories",
-    "!tsconfig.dist.json"
-  ],
-  "author": "",
-  "private": false,
-  "license": "MIT",
-  "dependencies": {
-    "@kaizen/draft-card": "^1.2.4",
-    "@kaizen/draft-page-layout": "^1.0.0",
-    "classnames": "^2.2.6",
-    "resize-observer-polyfill": "^1.5.1"
-  },
-  "peerDependencies": {
-    "@kaizen/design-tokens": "^2",
-    "react": "^16.9.0"
-  }
-}
