--- draft-packages/skirt/tsconfig.dist.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.dist.json",
-  "compilerOptions": {
-    "allowJs": false,
-    "declaration": true,
-    "noEmit": false,
-    "sourceMap": true
-  }
-}
