+++ draft-packages/stories/TitleBlockZen.stories.tsx
@@ -6,8 +6,7 @@ import starIcon from "@kaizen/component-library/icons/star-on.icon.svg"
 import reportSharingIcon from "@kaizen/component-library/icons/report-sharing.icon.svg"
 import arrowForwardIcon from "@kaizen/component-library/icons/arrow-forward.icon.svg"
 import { assetUrl } from "@kaizen/hosted-assets"
-import { Skirt, SkirtCard } from "@kaizen/draft-skirt"
-import { Container, Content } from "@kaizen/draft-page-layout"
+import { Container, Content, Skirt, SkirtCard } from "@kaizen/draft-page-layout"
 import {
   NavigationTab,
   TitleBlockZen,
